+++ src/index.js
@@ -190,6 +190,8 @@ export default function commonjs ( options = {} ) {
 
 				commonjsModules.set( id, true );
 				return transformed;
+			}).catch(err => {
+				this.error(err, err.loc);
 			});
 		}
 	};